Rename routes file
