+++ src/app/app.config.ts
@@ -20,7 +20,7 @@ import { withFetch } from '@angular/common/http';
 import { provideDateFnsAdapter } from '@angular/material-date-fns-adapter';
 import { MAT_DATE_LOCALE } from '@angular/material/core';
 import { csrfInterceptor } from '@app/auth/services/csrf';
-import { routes, TemplatePageTitleStrategy } from './app.routes';
+import { routes, TemplatePageTitleStrategy } from './app-routes';
 
 import {
   DEFAULT_DATE_FORMATS,
